move telegram-mcp and user-group services out from the project
