--- src/user_group_db/__init__.py
@@ -1,1 +1,0 @@
-
